fix: avoid defining  variable twice
The user is able to overwrite these values. However the user will probably use the "define" keyword. Thus the mdp_dict will contain "define" and "Define" keywords. As mdp is case insensitive - it will complaint about the "define" keyword being assigned twice. Setting Define to define for consistency and to avoid this situation